Use "tail" instead of "dd" to skip over the header
ADB backup headers are always 4 lines, regardless of encryption. Instead
of using a fixed amount of bytes (which is also correct, indeed), it is
possible to skip over the 4 first lines. This is good to avoid having
"dd" output the transfer summary, which in turn makes the output of
"abtool list" suitable to be passed to other commands e.g. via pipes.